+++ packages/isaacscript-common/src/functions/cards.ts
@@ -1,5 +1,6 @@
 import { CardType, ItemConfigCardType } from "isaac-typescript-definitions";
 import { itemConfig } from "../core/cachedClasses";
+import { ITEM_CONFIG_CARD_TYPES_FOR_CARDS } from "../core/constants";
 import {
   FIRST_CARD_TYPE,
   LAST_VANILLA_CARD_TYPE,
@@ -9,60 +10,7 @@ import {
   DEFAULT_CARD_DESCRIPTION,
 } from "../objects/cardDescriptions";
 import { CARD_NAMES, DEFAULT_CARD_NAME } from "../objects/cardNames";
-import {
-  CARD_TYPE_TO_ITEM_CONFIG_CARD_TYPE,
-  DEFAULT_CARD_TYPE,
-} from "../objects/cardTypeToItemConfigCardType";
-import { getEnumValues } from "./enums";
-import { getRandomSeed } from "./rng";
-import { addSetsToSet, getRandomSetElement } from "./set";
 import { iRange } from "./utils";
-
-const ITEM_CONFIG_CARD_TYPE_TO_CARD_TYPE_MAP = new Map<
-  ItemConfigCardType,
-  Set<CardType>
->();
-
-/**
- * Contains all of the entries in the `CardType` enum with the following types:
- * - ItemConfigCardType.TAROT
- * - ItemConfigCardType.SUIT
- * - ItemConfigCardType.SPECIAL
- * - ItemConfigCardType.TAROT_REVERSE
- */
-const CARD_SET = new Set<CardType>();
-
-function lazyInitCardMapsSets() {
-  // The card type to cards map should be valid for every card type, so we initialize it with empty
-  // sets.
-  for (const cardType of getEnumValues(ItemConfigCardType)) {
-    ITEM_CONFIG_CARD_TYPE_TO_CARD_TYPE_MAP.set(cardType, new Set<CardType>());
-  }
-
-  for (const card of getVanillaCardTypes()) {
-    const itemConfigCardType = getItemConfigCardType(card);
-    const cardTypeSet =
-      ITEM_CONFIG_CARD_TYPE_TO_CARD_TYPE_MAP.get(itemConfigCardType);
-    if (cardTypeSet === undefined) {
-      error(
-        `Failed to get the card set for item config card type: ${itemConfigCardType}`,
-      );
-    }
-    cardTypeSet.add(card);
-  }
-
-  // i.e. everything except for:
-  // - ItemConfigCardType.RUNE
-  // - ItemConfigCardType.SPECIAL_OBJECT
-  // - ItemConfigCardType.MODDED
-  const cards = getCardTypesOfType(
-    ItemConfigCardType.TAROT,
-    ItemConfigCardType.SUIT,
-    ItemConfigCardType.SPECIAL,
-    ItemConfigCardType.TAROT_REVERSE,
-  );
-  addSetsToSet(CARD_SET, cards);
-}
 
 /**
  * Helper function to get a card description from a Card enum value.
@@ -111,100 +59,28 @@ export function getCardName(cardType: CardType): string {
 }
 
 /**
- * Helper function to get a set of card types matching the `ItemConfigCardType`.
+ * Helper function to get the item config card type of a particular card, rune, or object. For
+ * example, the item config card type of `CardType.FOOL` is equal to `ItemConfigCardType.TAROT`.
  *
- * This function is variadic, meaning that you can you can specify N card types to get a set
- * containing cards that match any of the specified types.
+ * Returns undefined if the provided card type was not valid.
  */
-export function getCardTypesOfType(
-  ...itemConfigCardTypes: ItemConfigCardType[]
-): Set<CardType> {
-  if (ITEM_CONFIG_CARD_TYPE_TO_CARD_TYPE_MAP.size === 0) {
-    lazyInitCardMapsSets();
+export function getItemConfigCardType(
+  cardType: CardType,
+): ItemConfigCardType | undefined {
+  const itemConfigCard = itemConfig.GetCard(cardType);
+  if (itemConfigCard === undefined) {
+    return undefined;
   }
 
-  const matchingCardTypes = new Set<CardType>();
-  for (const itemConfigCardType of itemConfigCardTypes) {
-    const cardTypeSet =
-      ITEM_CONFIG_CARD_TYPE_TO_CARD_TYPE_MAP.get(itemConfigCardType);
-    if (cardTypeSet === undefined) {
-      error(
-        `Failed to get the card type set for item config type: ${itemConfigCardType}`,
-      );
-    }
-
-    for (const cardType of cardTypeSet.values()) {
-      matchingCardTypes.add(cardType);
-    }
-  }
-
-  return matchingCardTypes;
-}
-
-export function getItemConfigCardType(cardType: CardType): ItemConfigCardType {
-  const itemConfigCardType = CARD_TYPE_TO_ITEM_CONFIG_CARD_TYPE[cardType];
-
-  // Handle modded cards.
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  return itemConfigCardType === undefined
-    ? DEFAULT_CARD_TYPE
-    : itemConfigCardType;
+  return itemConfigCard.CardType;
 }
 
 /**
- * Has an equal chance of returning any card (e.g. Fool, Reverse Fool, Wild Card, etc.).
+ * Helper function to get an array with every valid vanilla card sub-type.
  *
- * This will not return:
- * - any runes
- * - any objects like Dice Shard
- * - any modded cards (since there is not a way to distinguish between modded cards and modded
- *   runes/objects)
- *
- * @param seedOrRNG Optional. The `Seed` or `RNG` object to use. If an `RNG` object is provided, the
- *                  `RNG.Next` method will be called. Default is `getRandomSeed()`.
- * @param exceptions Optional. An array of cards to not select.
+ * Note that unlike collectibles and trinkets, there are no gaps in the card types, so this is a
+ * simple range from the first card type to the last vanilla card type.
  */
-export function getRandomCard(
-  seedOrRNG: Seed | RNG = getRandomSeed(),
-  exceptions: CardType[] = [],
-): CardType {
-  return getRandomSetElement(CARD_SET, seedOrRNG, exceptions);
-}
-
-/**
- * @param itemConfigCardType The item config card type that represents the pool of cards to select
- *                           from.
- * @param seedOrRNG Optional. The `Seed` or `RNG` object to use. If an `RNG` object is provided, the
- *                  `RNG.Next` method will be called. Default is `getRandomSeed()`.
- * @param exceptions Optional. An array of cards to not select.
- */
-export function getRandomCardTypeOfType(
-  itemConfigCardType: ItemConfigCardType,
-  seedOrRNG: Seed | RNG = getRandomSeed(),
-  exceptions: CardType[] = [],
-): CardType {
-  const cardTypeSet = getCardTypesOfType(itemConfigCardType);
-  return getRandomSetElement(cardTypeSet, seedOrRNG, exceptions);
-}
-
-/**
- * Has an equal chance of returning any rune (e.g. Rune of Hagalaz, Blank Rune, Black Rune, Soul of
- * Isaac, etc.). This will never return a Rune Shard.
- *
- * @param seedOrRNG Optional. The `Seed` or `RNG` object to use. If an `RNG` object is provided, the
- *                  `RNG.Next` method will be called. Default is `getRandomSeed()`.
- * @param exceptions Optional. An array of runes to not select.
- */
-export function getRandomRune(
-  seedOrRNG: Seed | RNG = getRandomSeed(),
-  exceptions: CardType[] = [],
-): CardType {
-  const runesSet = getCardTypesOfType(ItemConfigCardType.RUNE);
-  runesSet.delete(CardType.RUNE_SHARD);
-  return getRandomSetElement(runesSet, seedOrRNG, exceptions);
-}
-
-/** Helper function to get an array with every valid vanilla card sub-type. */
 export function getVanillaCardTypes(): CardType[] {
   return iRange(FIRST_CARD_TYPE, LAST_VANILLA_CARD_TYPE);
 }
@@ -217,7 +93,12 @@ export function getVanillaCardTypes(): CardType[] {
  * - CardType.TAROT_REVERSE
  */
 export function isCard(cardType: CardType): boolean {
-  return CARD_SET.has(cardType);
+  const itemConfigCardType = getItemConfigCardType(cardType);
+  if (itemConfigCardType === undefined) {
+    return false;
+  }
+
+  return ITEM_CONFIG_CARD_TYPES_FOR_CARDS.has(itemConfigCardType);
 }
 
 /** Returns whether or not the given card type matches the specified item config card type. */